Add admin account when initializing
When the first user initializes the application, they need to give the organisation information and an account which will be able to configure the application later (an admin account).

We need to update the server and the UI to do so.

We introduce password field for the user to correctly set its password with indication of strength.